small tracemapper refactoring #12

+++ tools/route/tracemapper.py
@@ -46,9 +46,9 @@ def readFCD(traceFile, net, geo):
             trace = []
             last = v.id
         if geo:
-            trace.append(net.convertLonLat2XY(v.x, v.y))
+            trace.append(net.convertLonLat2XY(float(v.x), float(v.y)))
         else:
-            trace.append((v.x, v.y))
+            trace.append((float(v.x), float(v.y)))
     if trace:
         yield last, trace
 
@@ -83,7 +83,7 @@ if __name__ == "__main__":
                          help="generate polygon output for the mapped edges", metavar="FILE")
     optParser.add_option("--geo", action="store_true",
                          default=False, help="read trace with geo-coordinates")
-    optParser.add_option("--fill-gaps", default=0, type=int,
+    optParser.add_option("--fill-gaps", default=0., type=float,
                          help="repair disconnected routes bridging gaps of up to x meters")
     optParser.add_option("-g", "--gap-penalty", default=-1, type="float",
                          help="penalty to add for disconnected routes " +
